🚚(frontend) move dashboardButton in styled folder
DashboardButton are just styled Button and therefore should be in the dedicated
folder.
